enabled windows notifications, moved code to module
